requirements.txt
streamlit==1.31.1
groq==0.4.2
PyPDF2==3.0.1
chromadb==0.4.22
sentence-transformers==2.5.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.24 